Add clang Tools
This is a Linux based image so we can use the go clang integration to
use msan to check the sanitation of memory use within go code.

With the clang tools installed it should be possible to run:
  go test -msan
and expect to see if our Unit Test cause anything terrible to happen
with memory management.